+++ internal/services/ollama.go
@@ -7,6 +7,7 @@ import (
 	"iter"
 	"net/http"
 	"net/url"
+	"slices"
 
 	"github.com/MegaGrindStone/mcp-web-ui/internal/models"
 	"github.com/ollama/ollama/api"
@@ -41,15 +42,23 @@ func NewOllama(host, model string) Ollama {
 // for cancellation and a slice of messages representing the conversation history. The function returns
 // an iterator that yields response chunks as strings and potential errors. The response is streamed
 // incrementally, allowing for real-time processing of model outputs.
-func (o Ollama) Chat(ctx context.Context, messages []models.Message) iter.Seq2[string, error] {
-	return func(yield func(string, error) bool) {
+func (o Ollama) Chat(
+	ctx context.Context,
+	systemMessage string,
+	messages []models.Message,
+) iter.Seq2[models.Content, error] {
+	return func(yield func(models.Content, error) bool) {
 		msgs := make([]api.Message, len(messages))
 		for i, msg := range messages {
 			msgs[i] = api.Message{
 				Role:    msg.Role,
-				Content: msg.Content,
+				Content: models.RenderContents(msg.Contents),
 			}
 		}
+		msgs = slices.Insert(msgs, 0, api.Message{
+			Role:    "system",
+			Content: systemMessage,
+		})
 
 		t := true
 		req := api.ChatRequest{
@@ -62,7 +71,10 @@ func (o Ollama) Chat(ctx context.Context, messages []models.Message) iter.Seq2[s
 		defer cancel()
 
 		if err := o.client.Chat(ctx, &req, func(res api.ChatResponse) error {
-			if !yield(res.Message.Content, nil) {
+			if !yield(models.Content{
+				Type: models.ContentTypeText,
+				Text: res.Message.Content,
+			}, nil) {
 				cancel()
 			}
 			return nil
@@ -70,7 +82,7 @@ func (o Ollama) Chat(ctx context.Context, messages []models.Message) iter.Seq2[s
 			if errors.Is(err, context.Canceled) {
 				return
 			}
-			yield("", fmt.Errorf("error sending request: %w", err))
+			yield(models.Content{}, fmt.Errorf("error sending request: %w", err))
 		}
 	}
 }